feat: add material lists and task filters
