+++ src/app/dashboard/layout.tsx
@@ -5,7 +5,10 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
   return (
     <div className="min-h-screen flex flex-col">
       <header className="p-4 border-b flex justify-between items-center">
-        <h2 className="font-semibold text-xl">Mis Proyectos</h2>
+        <div className="flex items-center gap-4">
+          <h2 className="font-semibold text-xl">Mis Proyectos</h2>
+          <a href="/dashboard/tareas" className="text-sm text-blue-600 hover:underline">Mis tareas</a>
+        </div>
         <UserButton afterSignOutUrl="/" />
       </header>
       <main className="p-6">{children}</main>
